qcheck-lin index

Library qcheck-lin

This library exposes the following toplevel modules:

Lin

Lin_api The Lin_api module allows the user to describe the type signature of a tested module interface using a DSL of type combinators.